refactor(ollama-tui-test): use shared MCP executor
