+++ crates/llm/Cargo.toml
@@ -10,7 +10,8 @@ async-trait = "0.1.88"
 clap = { version = "4.5.43", features = ["derive"] }
 gemini-rs = { git = "https://github.com/dstoc/gemini-rs", branch = "include-thoughts", version = "2.0.0" }
 ollama-rs = { git = "https://github.com/dstoc/ollama-rs", branch = "RobJellinghaus/streaming-tools", version = "0.3.2", features = ["macros", "stream"] }
-rmcp = { version = "0.4.0", features = ["client"] }
+rmcp = { version = "0.4.0", features = ["client", "transport-child-process"] }
+serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.142"
-tokio = { version = "1.47.1", features = ["macros", "rt", "sync"] }
+tokio = { version = "1.47.1", features = ["fs", "macros", "process", "rt", "sync"] }
 tokio-stream = "0.1.17"
