+++ crates/ollama-tui-test/AGENTS.md
@@ -12,11 +12,6 @@ Terminal chat interface to Ollama with MCP tool integration.
 - ratatui
   - terminal UI
     - input handling via `crossterm`
-- rmcp
-  - connect to MCP servers
-- serde
-  - load MCP server configuration
-    - `serde_json` handles JSON parsing
 - textwrap
   - wrap chat history
 - termimad
