+++ crates/ollama-tui-test/Cargo.toml
@@ -10,15 +10,11 @@ tokio = { version = "1", features = ["full"] }
 tokio-stream = "0.1"
 ratatui = "0.29"
 crossterm = { version = "0.29", features = ["event-stream"] }
-rmcp = { version = "0.4", features = ["transport-child-process", "client"] }
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
 textwrap = "0.16"
 termimad = "0.33.0"
 tui-input = { version = "0.14.0", default-features = false, features = ["crossterm"] }
 unicode-width = "0.2.0"
 llm = { version = "0.1.0", path = "../llm" }
-async-trait = "0.1.88"
 
 [dev-dependencies]
 insta = "1.43.1"
